add huaweicloud plugin snat-rule action
